// Qanat.Web/src/app/shared/generated/model/water-account-contact-upsert-dto.ts
export class WaterAccountContactUpsertDto { 
    ContactName: string;
    ContactEmail?: string | null;
    ContactPhoneNumber?: string | null;
    Address: string;
    SecondaryAddress?: string | null;
    City: string;
    State: string;
    ZipCode: string;
    PrefersPhysicalCommunication?: boolean | null;
    AddressValidated?: boolean;
    AddressValidationJson?: string | null;
    WaterAccountID?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface WaterAccountContactUpsertDtoForm { 
    ContactName: FormControl<string>;
    ContactEmail?: FormControl<string>;
    ContactPhoneNumber?: FormControl<string>;
    Address: FormControl<string>;
    SecondaryAddress?: FormControl<string>;
    City: FormControl<string>;
    State: FormControl<string>;
    ZipCode: FormControl<string>;
    PrefersPhysicalCommunication?: FormControl<boolean>;
    AddressValidated?: FormControl<boolean>;
    AddressValidationJson?: FormControl<string>;
    WaterAccountID?: FormControl<number>;
}